requirements.txt
django
djangorestframework
python-decouple
whitenoise
gunicorn
djangorestframework-simplejwt
pika
django-dramatiq
pytest-django
pytest-cov

